Refactor hab evaluator
